pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "gm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "langchain-core>=1.2.7",
    "langgraph>=1.0.6",
    "pydantic-settings>=2.12.0",
    "tenacity>=9.1.2",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "respx>=0.22.0",
]


[tool.setuptools]
package-dir = { "" = "src" }
packages = ["gm"]


[tool.uv]
index-url = "https://pypi.org/simple"


[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]


[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["gm"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_untyped_defs = false
ignore_missing_imports = true
